Remove the camera delay
This speeds up image processing. The multiplier was hard to get fine tuned. When teams start, the robot can see over 13 tokens, which slows things down massively! The calculations and processing are still _slightly_ intense so should delay things enough for us to be comfortable with it being realistic enough