Extract the timestep thread to a legacy class
This avoids a variety of timing issues which introduce randomness
into the robot behaviours that we don't want.

At least the camera needs to have a time step explicitly forced
before it can provide useful information, so we handle that in
the Camera itself. Simple testing suggests that other sensors
may not be so affected, though it is hard to know whether that
is a natural consequence of them only being used after several
timesteps have been done as a matter of course.