Preserve the legacy behaviour as the default
It turns out that it's fairly easy to write code which doesn't
explicitly advance time and instead ends up in an infinite loop
waiting for time to advance on its own.

Given that of the two competitor code samples we have one managed
to hit this case in several places, we should aim to keep
compatibility where we can while providing a way out for teams
that want the alternative.