Implement wait_start via Webots customData field
This is closer to how we really run matches and removes the need
to reset the simulation as part of the setup process.

In turn, this will make it possible to automate video capture of
the matches without running into Webots limitations.

For the moment I've left the animation capture as starting at the
start of the match time, however it could now start earlier (i.e:
including the setup phase too) if we wanted.

Fixes https://github.com/srobo/competition-simulator/issues/200