Introduce environment variables for output control
This introduces RECORDING_STEM and LOG_FILENAME for controlling the
match recordings and robot logs respectively. It also removes use
of the 'match.txt' file for controlling filenames now that arbitrary
names can be passed in.

This actually breaks prepare-comp-match as it can no longer specify
the match, however that's fixable by re-working it to run the whole
match rather than just prepare the match.