First pass at full match automation
I've not yet worked out where the match file and recordings should
be put, but that's fairly easy now that this is all hooked up.

I had been expecting to need to use the data from the match file
to do the team log collation, yet of course we actually already
have the relevant information to hand in the runner. This may
mean that we should re-evaluate the way that that file works,
though I suspect that as-is is fine given that its eventual format
is roughly what it needs to be anyway for SRComp.