Fix mass ratio warnings
The extreme mass ratio between the wheel spacer and robot body causes a physics warning to be raised.
SInce the spacer is not attached to the axle and does not rotate it's mass won't effect the movement of the wheel.
This changes from using an arbitrary density to a fixed mass of 1 gram for each spacer