Use runpy rather than exec
The runpy package is what implements `python -m X.py` and `python X.py`
running, so it's pretty much designed for what we're trying to do here.

The main reason to switch to it is that it preserves the original file
name of the code being run and thus plays much better with debuggers.