Convert vision API to using radians
The SR vision API now uses radians due to moving to `zoloto`.
This commit does not change the shape of the API, just the units,
so that the API shape changes can happen separately.

This converts all computation and APIs, however keeps degrees in
the tests and other places where humans need to eyeball the numbers
to understand them.