Move the marker boundary points inwards by one marker pixel
Each marker pixel is 10% of the marker's size using the marker set
we use, with the outermost pixel value forming the white boundary
around the marker. Testing of our vision system indicates that it
doesn't need this boundary to be in shot for the detection to work,
so for detection purposes we can consider the outer points of the
marker to be the area bounded by that spacing pixel rather than
including it.