Update our promised libraries to recent versions
This matches the versions which will be on the kits, including
supporting packages we pin for clarity. Bumping these also aids
supporting Python 3.11.

Note also that numpy 1.25 no longer supports Python 3.8, so we're
forced to stay on 1.24 for the moment. (Thankfully that version
does support Python 3.11, so we're not bounded on that side).