Move to importlib.metadata
pkg_resources isn't part of the standard library an is in fact
deprecated. Using importlib.metadata is the expected replacement,
and is part of the standard library. This resolves a warning that
users would otherwise see when running Webots, which is likely to
confuse first-time users.

Fixes https://github.com/srobo/competition-simulator/issues/398