Introduce integration tests for our vision stack
The vision stack has been the hardest to keep working as well as
the one which has been refactored the most. This test harness aims
to make it easier to have confidence that the vision system is
working as intended. If only it were possible to do this for the
physical kits too...

As part of this we also drop the legacy tests which are known to
not be working. They'll be replaced shortly by integration tests.