Update the arena & token markers for vision axes reworking
This realigns the actual arena & token markers for the changes made
to the underlying marker object. I've manually tested with the
keyboard robot that these produce sensible values. The tops/sides
of the tokens aren't specifically arranged, however the markers
are I believe all correctly facing outward from the tokens.